+++ src/models/maskAssignation.ts
