+++ src/models/pserDetails.ts
