+++ src/models/trrade.ts
